-,VectorSpaceModel,VectorSpaceModel-method {wordVectors}	R Documentation

VectorSpaceModel subtraction

Description

Keep the VSM class when doing subtraction operations; make it possible to subtract a single row from an entire model.

Usage

## S4 method for signature 'VectorSpaceModel,VectorSpaceModel'
e1 - e2

Arguments

e1
A vector space model

e2
A vector space model of equal size OR a vector space model of a single row. If the latter (which is more likely) the specified row will be subtracted from each row.

Value

A VectorSpaceModel of the same dimensions and rownames as e1

I believe this is necessary, but honestly am not sure.

[Package wordVectors version 2.0 Index]