VectorSpaceModel-class {wordVectors}	R Documentation

Vector Space Model class

Description

A class for describing and accessing Vector Space Models like Word2Vec. The base object is simply a matrix with columns describing dimensions and unique rownames as the names of vectors. This package gives a number of convenience functions for printing and, most importantly, accessing these objects.

Value

An object of class "VectorSpaceModel"

Slots

magnitudes

The cached sum-of-squares for each row in the matrix. Can be cached to speed up similarity calculations

[Package wordVectors version 2.0 Index]