demo_vectors {wordVectors}	R Documentation

999 vectors trained on teaching evaluations

Description

A sample VectorSpaceModel object trained on about 15 million teaching evaluations, limited to the 999 most common words. Included for demonstration purposes only: there's only so much you can do with a 999 length vocabulary.

Usage

demo_vectors

Format

A VectorSpaceModel object of 999 words and 500 vectors

Details

You're best off downloading a real model to work with, such as the precompiled vectors distributed by Google at https://code.google.com/archive/p/word2vec/

Source

Trained by package author.

[Package wordVectors version 2.0 Index]