project {wordVectors}	R Documentation

Project each row of an input matrix along a vector.

Description

As with 'cosineSimilarity

Usage

project(matrix, vector)

Arguments

matrix
A matrix or VectorSpaceModel

vector
A vector (or object coercable to a vector) of the same length as the VectorSpaceModel.

Value

A new matrix or VectorSpaceModel of the same dimensions as 'matrix', each row of which is parallel to vector.

If the input is a matrix, the output will be a matrix: if a VectorSpaceModel, it will be a VectorSpaceModel.

[Package wordVectors version 2.0 Index]