square_magnitudes {wordVectors}	R Documentation

Square Magnitudes with caching

Description

square_magnitudes Returns the square magnitudes and caches them if necessary

Usage

square_magnitudes(object)

Arguments

VectorSpaceModel
A matrix or VectorSpaceModel object

Value

A vector of the square magnitudes for each row

[Package wordVectors version 2.0 Index]