write.binary.word2vec {wordVectors}	R Documentation

Write in word2vec binary format

Description

Write in word2vec binary format

Usage

write.binary.word2vec(model, filename)

Arguments

model
The wordVectors model you wish to save. (This can actually be any matrix with rownames, if you want a smaller binary serialization in single-precision floats.)

filename
The file to save the vectors to. I recommend ".bin" as a suffix.

Value

Nothing

[Package wordVectors version 2.0 Index]